+++ src/components/errorMessage/errorMessage.jsx
@@ -1,0 +1,8 @@
+import img from "./unlike.svg"
+
+const ErrorMessage=()=>{
+    return (
+        <img src={img} style={{width:"250px", height:"250px", margin: "0 auto"}} alt="ddd"/>
+    )
+}
+export default ErrorMessage ;
