--- src/components/randomChar/RandomChar.js
@@ -1,79 +1,0 @@
-import { Component } from "react";
-import MarvelService from "../../services/MarvelService";
-
-import "./randomChar.scss";
-import thor from "../../resources/img/thor.jpeg";
-import mjolnir from "../../resources/img/mjolnir.png";
-import { render } from "@testing-library/react";
-import Spinner from "../spinner/Spinner";
-
-class RandomChar extends Component {
-  constructor(props) {
-    super(props);
-    this.updateChar();
-  }
-  state = {
-    char: {},
-    loading:true
-  };
-
-  marvelService = new MarvelService();
-
-  onCharLoaded =(char)=>{
-    this.setState({char, loading:false})
-  }
-
-  updateChar = () => {
-    const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
-    this.marvelService.getCharacter(id).then(this.onCharLoaded);
-  };
-
-  render() {
-    const { char ,loading} = this.state;
-
-
-
-    return (
-      <div className="randomchar">
-      {loading? <Spinner/>: <View char={char}/>}
-        <div className="randomchar__static">
-          <p className="randomchar__title">
-            Random character for today!
-            <br />
-            Do you want to get to know him better?
-          </p>
-          <p className="randomchar__title">Or choose another one</p>
-          <button className="button button__main">
-            <div className="inner">try it</div>
-          </button>
-          <img src={mjolnir} alt="mjolnir" className="randomchar__decoration" />
-        </div>
-      </div>
-    );
-  }
-}
-
-const View = ({char})=>{
-    const {name,description,thumbnail,homepage,wiki} = char;
-    return(  <div className="randomchar__block">
-    <img
-      src={thumbnail}
-      alt="Random character"
-      className="randomchar__img"
-    />  
-    <div className="randomchar__info">
-      <p className="randomchar__name">{name}</p>
-      <p className="randomchar__descr">{ description?description.substr(0,300)+"...":"Нету описания"}</p>
-      <div className="randomchar__btns">
-        <a href={homepage} className="button button__main">
-          <div className="inner">homepage</div>
-        </a>
-        <a href={wiki} className="button button__secondary">
-          <div className="inner">Wiki</div>
-        </a>
-      </div>
-    </div>
-  </div>)
-}
-
-export default RandomChar;
